% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_misc.R
\name{balanced.cv.fold}
\alias{balanced.cv.fold}
\title{Split a dataset for Cross Validation taking into account class balance}
\usage{
balanced.cv.fold(y, num.cv = 10)
}
\arguments{
\item{y}{the class labels of each sample of the dataset}

\item{num.cv}{number of cross validation required}
}
\value{
a factor of num.cv levels that assign to each sample a test fold
}
\description{
Split a dataset for Cross Validation taking into account class balance
}
