% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_misc.R
\name{bhattacharyya.coefficient}
\alias{bhattacharyya.coefficient}
\title{Compute Bhattacharyya coefficient needed for Hellinger distance}
\usage{
bhattacharyya.coefficient(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\value{
a square matrix containing Bhattacharyya coefficient for each pair of row in x
}
\description{
Compute Bhattacharyya coefficient needed for Hellinger distance
}
\author{
Julien Prados
}
