% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrm2.R
\name{bmrm2}
\alias{bmrm2}
\title{Bundle Methods for Regularized Risk Minimization}
\usage{
bmrm2(riskFun, LAMBDA = 1, MAX_ITER = 100, EPSILON_TOL = 0.01, w0 = 0,
  verbose = TRUE)
}
\arguments{
\item{riskFun}{the loss function to use in the optimization (e.g.: hingeLoss, softMarginVectorLoss). 
The function must evaluate the loss value and its gradient for a given point vector (w).}

\item{LAMBDA}{control the regularization strength in the optimization process. This is the value used as coefficient of the regularization term.}

\item{MAX_ITER}{the maximum number of iteration to perform. The function stop with a warning message if the number of iteration exceed this value}

\item{EPSILON_TOL}{control optimization stoping criteria: the optimization end when the optimization gap is below this threshold}

\item{w0}{initial weight vector where optimization start}

\item{verbose}{a length one logical. Show progression of the convergence on stdout}
}
\value{
the optimized weight vector, with attribute "log" beging a data.frame storing a trace of important values of the optimization process.
}
\description{
Implement Bundle Methods for Regularized Risk Minimization as described in Teo et. al 2007.
Find w that minimize: LAMBDA*regularization_norm(w) + lossfun(w)
where regularization_norm is L1 (for L2, use nrbm()).
}
\references{
Teo et al.
  A Scalable Modular Convex Solver for Regularized Risk Minimization.
  KDD 2007
}
\seealso{
\code{\link{hingeLoss}} \code{\link{softMarginVectorLoss}}
}
\author{
Julien Prados
}
