% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scalarRegressionLosses.R
\name{epsilonInsensitiveRegressionLoss}
\alias{epsilonInsensitiveRegressionLoss}
\title{The loss function to perform a epsilon-insensitive regression (Vapnik et al. 1997)}
\usage{
epsilonInsensitiveRegressionLoss(x, y, epsilon)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values representing the training labels for each instance in x}

\item{epsilon}{a numeric value setting tolerance of the epsilon-regression}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to perform a epsilon-insensitive regression (Vapnik et al. 1997)
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

