% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scalarClassificationLosses.R
\name{fbetaLoss}
\alias{fbetaLoss}
\title{F beta score loss function}
\usage{
fbetaLoss(x, y, beta = 1)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values in (-1,+1) representing the training labels for each instance in x}

\item{beta}{a numeric value setting the beta parameter is the f-beta score}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
F beta score loss function
}
\references{
Teo et al.
  A Scalable Modular Convex Solver for Regularized Risk Minimization.
  KDD 2007
}
\seealso{
bmrm
}

