% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_loss.R
\name{gradient}
\alias{gradient}
\alias{gradient.default}
\alias{gradient<-}
\alias{gradient<-.default}
\title{Return or set gradient attribute}
\usage{
gradient(x, ...)

\method{gradient}{default}(x, ...)

gradient(x, ...) <- value

\method{gradient}{default}(x, ...) <- value
}
\arguments{
\item{x}{any R object}

\item{...}{additional paramters}

\item{value}{new gradient value to set}
}
\value{
attr(x,"gradient")
}
\description{
Return or set gradient attribute
}
\details{
gradient attribute isused by loss/risk function to return the gradient of
the function at a given point together with the function value
}

