% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_misc.R
\name{hellinger.dist}
\alias{hellinger.dist}
\title{Compute Hellinger distance}
\usage{
hellinger.dist(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\value{
an object of class "dist" with Hellinger distance of each pair of row in x
}
\description{
Compute Hellinger distance
}
\author{
Julien Prados
}
