% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scalarRegressionLosses.R
\name{ladRegressionLoss}
\alias{ladRegressionLoss}
\title{The loss function to perform a least absolute deviation regression}
\usage{
ladRegressionLoss(x, y)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values representing the training labels for each instance in x}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to perform a least absolute deviation regression
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

