% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossScalarRegression.R
\name{logisticRegressionLoss}
\alias{logisticRegressionLoss}
\title{The loss function to perform a logistic regression}
\usage{
logisticRegressionLoss(x, y)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values representing the training labels for each instance in x}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to perform a logistic regression
}
\examples{
  x <- cbind(data.matrix(iris[1:2]),1)
  y <- c(-1,1,1)[iris$Species]
  w <- bmrm(logisticRegressionLoss(x,y),LAMBDA=1,verbose=TRUE)
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

