% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_core.R
\name{lvalue}
\alias{lvalue}
\alias{lvalue<-}
\alias{lvalue.default}
\alias{lvalue<-.default}
\title{Return or set lvalue attribute}
\usage{
lvalue(x, ...)

\method{lvalue}{default}(x, ...)

lvalue(x, ...) <- value

\method{lvalue}{default}(x, ...) <- value
}
\arguments{
\item{x}{any R object}

\item{...}{additional paramters}

\item{value}{new loss value to set}
}
\value{
attr(x,"lvalue")
}
\description{
Return or set lvalue attribute
}
\details{
lvalue attribute is used by loss/risk function to return the loss value of
the function at a given point together with the function gradient
}
