% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmc.R
\name{predict.mmc}
\alias{predict.mmc}
\title{Predict class of new instances according to a mmc model}
\usage{
\method{predict}{mmc}(object, x, ...)
}
\arguments{
\item{object}{a mmc object}

\item{x}{a matrix similar to the dataset used, i.e. where rows are instances for
which class must be predicted}

\item{...}{unused, present to satisfy the generic predict() prototype}
}
\value{
a integer vector whose length match nrow(x) and containing the predicted 
        class for each of the given instances.
}
\description{
Predict class of new instances according to a mmc model
}
