% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmLP.R
\name{predict.svmLP}
\alias{predict.svmLP}
\title{Predict function for svmLP models}
\usage{
\method{predict}{svmLP}(object, x, ...)
}
\arguments{
\item{object}{an object of class svmLP}

\item{x}{data.matrix to predict}

\item{...}{unused, present to satisfy the generic predict() prototype}
}
\value{
prediction for row of x, with an attribute "decision.value"
}
\description{
Predict function for svmLP models
}
\author{
Julien Prados
}

