% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_misc.R
\name{print.roc.stat}
\alias{print.roc.stat}
\title{Generic method overlad to print object of class roc.stat}
\usage{
\method{print}{roc.stat}(x, ...)
}
\arguments{
\item{x}{a roc.stat object return by the function roc.stat}

\item{...}{additional parameters}
}
\description{
Generic method overlad to print object of class roc.stat
}
