% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_misc.R
\name{rank.linear.weights}
\alias{rank.linear.weights}
\title{Rank linear weight of a linear model}
\usage{
rank.linear.weights(w)
}
\arguments{
\item{w}{a numeric vector of linear weights}
}
\value{
a data.frame with a rank for each feature as well as z-score, p-value, and false discovery rate.
}
\description{
Rank linear weight of a linear model
}
