% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_ML.R
\name{roc.curve}
\alias{roc.curve}
\title{Compute data for ROC curve plotting}
\usage{
roc.curve(f, y)
}
\arguments{
\item{f}{decision value for each instance}

\item{y}{a logical that specify binary labels}
}
\value{
a 3 columns data.frame() with 'FPR' and 'TPR' at each threshold value 'f'
}
\description{
Compute data for ROC curve plotting
}
\author{
Julien Prados, inspired by Bob Horton code
}

