% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rowmean}
\alias{rowmean}
\title{Columun means of a matrix based on a grouping variable}
\usage{
rowmean(x, group, ...)
}
\arguments{
\item{x}{a matrix}

\item{group}{a factor with one element per row of x}

\item{...}{additional arguments are passed to rowsum()}
}
\value{
a matrix containing the means, with one row per level of group.
}
\description{
Similar to rowsum, but for mean values.
}
