% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmLP.R
\name{svmMulticlassLP}
\alias{svmMulticlassLP}
\title{SVM Linear Program}
\usage{
svmMulticlassLP(x, y, LAMBDA = 1, instance.weights = 1)
}
\arguments{
\item{x}{a numeric data matrix.}

\item{y}{a response factor of >=2 levels for each row of x.}

\item{LAMBDA}{control the regularization strength in the optimization process. This is the value used as coefficient of the regularization term.}

\item{instance.weights}{a numeric vector of weight for each row of x, recycled as necessary.}
}
\value{
the optimized weight matrix
}
\description{
Solve a linear program implementing multiclass-SVM 
with L1 regularization and L1 loss. It solves:
min_w LAMBDA*|w| + sum_i(e_i);
s.t. <w.x_i> - <w.x_j> >= 1-e_i; e_i >= 0
where |w| is the L1-norm of w
}
\examples{
  x <- cbind(100,data.matrix(iris[1:4]))
  y <- iris$Species
  w <- svmMulticlassLP(x,y)
  table(predict(w,x),y)
}
\author{
Julien Prados
}

