% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossScalarRegression.R
\name{epsilonInsensitiveRegressionLoss}
\alias{epsilonInsensitiveRegressionLoss}
\title{The loss function to perform a epsilon-insensitive regression (Vapnik et al. 1997)}
\usage{
epsilonInsensitiveRegressionLoss(x, y, epsilon)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values representing the training labels for each instance in x}

\item{epsilon}{a numeric value setting tolerance of the epsilon-regression}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to perform a epsilon-insensitive regression (Vapnik et al. 1997)
}
\examples{
  x <- cbind(data.matrix(iris[1:2]),1)
  y <- iris[[3]]
  w <- bmrm(epsilonInsensitiveRegressionLoss(x,y,1),LAMBDA=0.1,verbose=TRUE)
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

