% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossScalarRegression.R
\name{ladRegressionLoss}
\alias{ladRegressionLoss}
\title{The loss function to perform a least absolute deviation regression}
\usage{
ladRegressionLoss(x, y)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values representing the training labels for each instance in x}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to perform a least absolute deviation regression
}
\examples{
  x <- cbind(data.matrix(iris[1:2]),1)
  y <- iris[[3]]
  w <- bmrm(ladRegressionLoss(x,y),LAMBDA=0.1,verbose=TRUE)
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

