% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{lmsRegressionLoss}
\alias{lmsRegressionLoss}
\title{The loss function to perform a least mean square regression}
\usage{
lmsRegressionLoss(x, y, loss.weights = 1)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values representing the training labels for each instance in x}

\item{loss.weights}{numeric vector of loss weights to incure for each instance of x. 
Vector length should match length(y), but values are cycled if not of identical size.}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to perform a least mean square regression
}
\examples{
  x <- cbind(data.matrix(iris[1:2]),1)
  y <- iris[[3]]
  w <- bmrm(lmsRegressionLoss(x,y),LAMBDA=0.1,verbose=TRUE)
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

