% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossVectorial.R
\name{ontologyLoss}
\alias{ontologyLoss}
\title{Ontology Loss Function}
\usage{
ontologyLoss(x, y, l = 1 - table(seq_along(y), y), dag = diag(nlevels(y)))
}
\arguments{
\item{x}{instance matrix, where x(t,) defines the features of instance t}

\item{y}{target vector where y(t) is an integer encoding target of x(t,)}

\item{l}{loss matrix. l(t,p(t)) must be the loss for predicting target p(t) instead of y(t) 
for instance t. By default, the parameter is set to a 0/1 loss matrix.}

\item{dag}{a numeric matrix defining the path in the Direct Acyclic Graph (DAG) to each class label}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
Ontology Loss Function
}
\examples{
  # -- Load the data
  x <- data.matrix(iris[1:2])
  y <- iris$Species
  dag <- matrix(c(1,0,0,0,
                  0,1,1,0,
                  0,1,0,1),3,4)
  w <- nrbm(ontologyLoss(x,y,dag=dag))
  w <- matrix(w,ncol(x))
  max.col(x \%*\% w \%*\% dag)
}
\references{
Teo et al.
  A Scalable Modular Convex Solver for Regularized Risk Minimization.
  KDD 2007
}

