% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossScalarRegression.R
\name{quantileRegressionLoss}
\alias{quantileRegressionLoss}
\title{The loss function to perform a quantile regression}
\usage{
quantileRegressionLoss(x, y, q = 0.5)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values representing the training labels for each instance in x}

\item{q}{a numeric value in the range [0-1] defining quantile value to consider}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to perform a quantile regression
}
\examples{
  x <- cbind(data.matrix(iris[1:2]),1)
  y <- iris[[3]]
  w <- bmrm(quantileRegressionLoss(x,y),LAMBDA=0.1,verbose=TRUE)
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

