% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossScalarClassification.R
\name{rocLoss}
\alias{rocLoss}
\title{The loss function to maximize area under the ROC curve}
\usage{
rocLoss(x, y)
}
\arguments{
\item{x}{matrix of training instances (one instance by row)}

\item{y}{numeric vector of values in (-1,+1) representing the training labels for each instance in x}
}
\value{
a function taking one argument w and computing the loss value and the gradient at point w
}
\description{
The loss function to maximize area under the ROC curve
}
\examples{
  x <- cbind(data.matrix(iris[1:2]),1)
  y <- c(-1,1,1)[iris$Species]
  w <- bmrm(rocLoss(x,y),LAMBDA=0.1,verbose=TRUE)
}
\references{
Teo et al.
  Bundle Methods for Regularized Risk Minimization
  JMLR 2010
}
\seealso{
bmrm
}

